requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
tensorflow
torch
torchvision
transformers
pillow